DSoundHelper.DefaultCaptureDevice has modopt(IsConstModifier).
Modifiers are considered part of the type, and therefore
references do not link correctly without it. C# doesn't provide
a way to declare modifiers.

All public readonly fields probably have this modifier, but I
haven't checked. I'm also not sure we should export the fields
until we have some idea of what they should be, since a missing
field reference is the easiest way to tell that they're being used.